fix: address self-upgrade review followups
